Use UCL instead of LZMA
Using `--lzma` on such a large binary is causing long loading times - not just when running, but also when examining the executable in the file explorer (potentially due to the virus scanner). With many of these compressed executables in the same folder, the CPU shoots up to 100% just examining them, before they are even run.

There is not a dramatic difference between UCL and LZMA, anyway - 1 MB is a decent price to pay for significantly faster loading times.