Use libaria2 instead of simple HTTP downloader
Now this will work with magnet links, ftp links, Google Drive hotlinks, you name it. Maybe torrents as well, but haven't tested it.

Also, this will only build with MinGW on Windows because aria2 is allergic to MSVC.

You will also need an OpenSSL root CA file to be placed in the same directory as the launcher:

https://github.com/q3aql/aria2-static-builds/blob/master/certs/ca-certificates.crt